Add Change Sections Module
