+++ index.js
@@ -1,4 +1,5 @@
 import BooksList from './modules/booksList.js';
+import { handleListBookClick, handleAddNewSectionClick, handleContactSectionClick } from './modules/changeSections.js';
 
 const booksList = new BooksList();
 const addButton = document.getElementById('add-book');
@@ -23,36 +24,6 @@ const options = {
 
 document.getElementById('date').innerHTML = currentDate.toLocaleDateString('en-US', options);
 
-const listBook = document.getElementById('listBooks');
-const addNewSection = document.getElementById('addNewBook');
-const contactSection = document.getElementById('contactSection');
-const sectionUno = document.getElementById('list');
-const sectionDos = document.getElementById('add-new');
-const sectionTres = document.getElementById('contact');
-
-listBook.addEventListener('click', () => {
-  sectionUno.style.display = 'block';
-  sectionDos.style.display = 'none';
-  sectionTres.style.display = 'none';
-  listBook.style.color = 'blue';
-  addNewSection.style.color = 'black';
-  contactSection.style.color = 'black';
-});
-
-addNewSection.addEventListener('click', () => {
-  sectionUno.style.display = 'none';
-  sectionDos.style.display = 'block';
-  sectionTres.style.display = 'none';
-  listBook.style.color = 'black';
-  addNewSection.style.color = 'blue';
-  contactSection.style.color = 'black';
-});
-
-contactSection.addEventListener('click', () => {
-  sectionUno.style.display = 'none';
-  sectionDos.style.display = 'none';
-  sectionTres.style.display = 'flex';
-  listBook.style.color = 'black';
-  addNewSection.style.color = 'black';
-  contactSection.style.color = 'blue';
-});
+listBook.addEventListener('click', handleListBookClick);
+addNewSection.addEventListener('click', handleAddNewSectionClick);
+contactSection.addEventListener('click', handleContactSectionClick);
